Fix keyboard support. Long tap support.
Fixed the keyboard support to work again. The copied in code was relying upon PostData working differently and was broken. As a result of this fix the newly imported code is working. That code fixes the ability to use 'backspace' and 'enter' keys for text input, which never worked before.

Altered home button to use wda pressButton function instead of 'homescreen', so that it can be used to dismiss 'launcher deletion mode'. It still also effectively works as a homescreen button.

Added support for "long tap".